Disable lens for the tabpage if any window in it is disabled.
The problem is that if a window is diabled, surely we don't resize any
window when we enter it. However, when we enter another window that is
not disabled, now the disabled window will be resized. And when we jump
back, it won't be resized back as it is disabled.